GC unused assert includes
