+++ searchcore/src/vespa/searchcore/proton/matching/partial_result.h
@@ -48,7 +48,7 @@ public:
         _sortData.push_back(sd);
         _sortDataSize += sd.second;
     }
-    virtual void merge(Source &rhs) override;
+    void merge(Source &rhs) override;
 };
 
 }
