+++ searchlib/src/vespa/searchlib/bitcompression/posocccompression.cpp
@@ -7,6 +7,7 @@
 #include <vespa/searchlib/index/postinglistparams.h>
 #include <vespa/vespalib/stllike/asciistream.h>
 #include <vespa/vespalib/data/fileheader.h>
+#include <cassert>
 
 #include <vespa/log/log.h>
 LOG_SETUP(".posocccompression");
