+++ searchlib/src/vespa/searchlib/test/diskindex/threelevelcountbuffers.cpp
@@ -1,6 +1,7 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "threelevelcountbuffers.h"
+#include <cassert>
 
 namespace search::diskindex {
