+++ vespalib/src/vespa/vespalib/geo/zcurve.cpp
@@ -10,15 +10,38 @@ namespace vespalib::geo {
 
 namespace {
 
+    /**
+ * An area defined by its upper left and lower right corners. The
+ * z-coordinates between these corners act as a spacial
+ * over-estimation of the actual area. These areas may never cross
+ * signed borders, since that would break the whole concept of
+ * hierarchical spatial partitioning.
+ **/
+struct Area {
+    const ZCurve::Point min;
+    const ZCurve::Point max;
+    Area(const Area &rhs) = default;
+    Area(int32_t min_x, int32_t min_y,
+         int32_t max_x, int32_t max_y)
+            : min(min_x, min_y), max(max_x, max_y)
+    {
+        assert((min_x <= max_x) && ((min_x < 0) == (max_x < 0)));
+        assert((min_y <= max_y) && ((min_y < 0) == (max_y < 0)));
+    }
+    Area &operator=(Area &&rhs) { new ((void*)this) Area(rhs); return *this; }
+    int64_t size() const { return (static_cast<int64_t>(max.x) - min.x + 1) * (static_cast<int64_t>(max.y) - min.y + 1); }
+    int64_t estimate() const { return (max.z - min.z + 1); }
+    int64_t error() const { return estimate() - size(); }
+};
+
 class ZAreaQueue
 {
 private:
     struct MaxAreaErrorCmp {
-        bool operator()(const ZCurve::Area &a, const ZCurve::Area &b) const {
+        bool operator()(const Area &a, const Area &b) const {
             return (a.error() > b.error());
         }
     };
-    using Area = ZCurve::Area;
     using Range = ZCurve::Range;
     using RangeVector = ZCurve::RangeVector;
     using Queue = PriorityQueue<Area, MaxAreaErrorCmp, LeftArrayHeap>;
@@ -61,7 +84,6 @@ public:
 class ZAreaSplitter
 {
 private:
-    using Area = ZCurve::Area;
     using RangeVector = ZCurve::RangeVector;
 
     ZAreaQueue _queue;
